feat: useResultStatus and ResultStatus type separated into two files and moved into common folder
