+++ src/services/global/LiveModel/ImportLM.ts
@@ -1,7 +1,7 @@
 import { Loading } from '../../../components/lit-components/loading-wrapper';
 import { Message } from '../../../components/lit-components/message-wrapper';
 import config from '../../../config';
-import useResultsStatus from '../../../stores/ComputationManager/useResultsStatus';
+import useResultsStatus from '../../../stores/ComputationManager/ResultStatus/useResultsStatus';
 import useVariablesStore from '../../../stores/LiveModel/useVariablesStore';
 import useTabsStore from '../../../stores/Navigation/useTabsStore';
 import { EdgeMonotonicity, type Variable } from '../../../types';
